\name{plot}

\docType{methods}

\alias{plot,envelope_model,missing-method}
\alias{plot,maxent_model,ANY-method}

\title{
Plot predictor values 
}

\description{
Plot predictor values for occurrence (presence and absence) data in a model object.
}

\usage{
\S4method{plot}{envelope_model,missing}(x, a = 1, b = 2, p = 0.9, ...)

\S4method{plot}{maxent_model,ANY}(x, y, ...)
}

\arguments{
\item{x}{model object}
\item{a}{describe}
\item{b}{describe}
\item{p}{describe}
\item{y}{missing}
\item{...}{arguments passed to \link[base]{plot}}
}


\keyword{spatial}

