\name{predicts-package}

\alias{predicts-package}
\alias{predicts}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package implements a few species distribution models, including an R link to the 'maxent' model, and native implementations of Bioclim and Domain. It also provides a number of functions that can assist in using Boosted Regresssion Trees.

A good place to start is the vignette, which you can access by typing \code{vignette('sdm', 'predicts')}

In addition there are a number of functions, such sampling background points, k-fold sampling, and for model evaluation (AUC) that are useful for these and for other species distribution modeling methods available in R (e.g. GLM, GAM, and RandomForest).
}

\author{
Robert J. Hijmans
}

\keyword{ package }
\keyword{ spatial }

