\name{RMSE}

\alias{RMSE}
\alias{RMSE_null}

\title{
Root Mean Square Error
}

\description{
Compute the Root Mean Square Error (RMSE)
}


\usage{
RMSE(obs, prd, na.rm=FALSE)

RMSE_null(obs, prd, na.rm=FALSE)
}
 
\arguments{
  \item{obs}{observed values}
  \item{prd}{predicted values}
  \item{na.rm}{logical. If \code{TRUE}, \code{NA}s are removed}
}

\value{
numeric
}

