\name{predicts-package}

\alias{predicts-package}
\alias{predicts}

\docType{package}

\title{
Spatial prediction
}

\description{
This package implements functions for spatial predictions methods, especially spatial (species) distribution models, including an R link to the 'maxent' model.
}

\author{
Robert J. Hijmans
}

\keyword{ package }
\keyword{ spatial }

